Start on network membership
This implements a `getMembers` method on `StarOverlayNetwork` that can
be called from the hub and from every spoke, and which returns a `Set`
containing every member belonging to that particular network.

Also, add a `onJoin(callback)` method which invokes the callback (on the
hub as well as on spokes) every time a new member joins the network.