Upgrade to google-protobuf 3.4.0 and work around Buffer incompatibility
Previously, we didn't even explicitly depend on google-protobuf and were
accidentally inheriting the dependency from tachyon. This fixes that.
Also, the new version of google-protobuf is now incompatibile with
Node's Buffer object because they do a direct comparison of the
.constructor property of the given object rather than an instanceof
check. Since Buffer is a subclass of Uint8Array, we just reassign the
.constructor property as Uint8Array to convince Google's type check to
allow it through. Kinda yucky but the only way I could find that didn't
involve copying the buffer.