Disconnect when ice connection state becomes 'closed'
Previously, we would detect a connection as closed only when its state
became `disconnected`. However, when a connection terminates abruptly
(e.g. when the OS goes into sleep mode and/or the network card is turned
off) Chromium puts the `iceConnectionState` straight into `closed` mode.

With this commit we are adding handling for the `closed` connection
state, which addresses the `Failed to execute 'send'` exception we were
observing.

I am not sure testing this behavior is feasible with simple unit tests,
so I believe we should defer this to the QA phase for now. If we find
this code path to be frequently subject to regressions, I think
investing more time on writing a more sophisticated integration test
could be justifiable.