Inject an authTokenProvider and get tokens just-in-time
We could encounter invalid token errors for a variety of reasons, such
as the user pasting garbage or the user explicitly deleting their auth
token on github.com. When this happens, we want to prompt the user for a
new token. This means the token isn't a constant, and injecting it
directly into various layers of the system is problematic when it comes
time to update it. Injecting a *provider*, which *is* constant, makes
updating the token as trivial as telling the provider to clear its cache
when we hit invalid token errors and requesting a new token.

/cc @jasonrudolph