Handle invalid token errors when sending offer/answer signals
We handle invalid tokens when offering by rejecting the connected
promise with the invalid token error.

We handle invalid token errors when answering by disconnecting and
dispatching onError listeners with the error, since there's no relevant
calling code on the stack when we receive a signal and answer.

In both cases, we tell the auth token provider to forget its token.