Introduce sign in and sign out concepts
With this commit, we are inverting the dependency to retrieve a token.
Instead of calling `authTokenProvider.getToken`, users of the API will
now need to call `signIn` and provide the token they want to
authenticate with.

This simplifies the UI interaction in the real-time package and allows
us to automatically sign out in the following scenarios:

* Answering a negotiation offer when the OAuth token is no longer valid
* Joining a portal when the OAuth token is no longer valid
* Creating a portal when the OAuth token is no longer valid

When that happens, the package will be notified with an `onSignOut`
event, which can be used to e.g. redirect users to the login dialog.

Signed-off-by: Jason Rudolph <jasonrudolph@github.com>