Toggle selections when tether changes after integrating a tether update
This is necessary because the tether of the site with the lowest id
could be retracted. If another site changes their tether and introduces
a cycle, we need to un-retract the tether of the new leader (i.e. the
site with the lowest id) and propagate its position to all the new
followers.