Update identity assertions to be more resilient to change
After upgrading to teletype-server@v0.17.0, several identity-related
tests began failing because they expected the identity object to only
contain a `login` property, and they didn't expect the inclusion of the
new `id` property. To fix these test failures and to make the tests more
resilient to potential future additive changes to the properties of the
identity object, this commit updates the tests to assert the value of
the `login` property, not the value of the entire identity object.