Don't call `portalDelegate.setActiveEditorProxy` if editor didn't change
This addresses a bug in the contract of
`PortalDelegate.setActiveEditorProxy`. When rapidly creating a new
editor proxy, switching to it, destroying it and then going back to the
previously active editor proxy, we would end up calling
`setActiveEditorProxy` on the delegate twice with the same value.

This violated the contract with the Teletype package (which provides an
implementation of a portal delegate), which was throwing the following
exception when the above scenario occurred:

> Uncaught (in promise) TypeError: Cannot read property 'getItems' of undefined
>     at GuestPortalBinding.replaceActivePaneItem (~/.atom/packages/teletype/lib/guest-portal-binding.js:159:25)
>     at lastSetActiveEditorProxyPromise.lastSetActiveEditorProxyPromise.then (~/.atom/packages/teletype/lib/guest-portal-binding.js:98:20)

Signed-off-by: Jason Rudolph <jasonrudolph@github.com>