Send disconnection message when calling PeerConnection.disconnect()
On Electron <= 1.6 we used to receive an `iceconnectionstatechange`
event whenever the other end of the connection hung up. On Electron 1.7
and newer this event is not being fired anymore, leaving one side of the
connection potentially open forever.

With this commit a disconnection message will be explicitly sent every
time one of the two sides of the connection decides to terminate the
communication. The other party can then interpret this message and, in
turn, close its side of the connection.