Bump protocol version from 5 to 6
Back in #48 we enhanced the first byte sent with every message in
PeerConnection to allow specifying if the message was a disconnection
message, a multipart message or a simple message.

Doing so, however, changed the protocol in a non-backward compatible
manner, as older clients are unable to interpret disconnection messages.
Hence, with this commit we are bumping the protocol version so that
everybody is forced to upgrade to the latest teletype version.