Introduce portal.addEditorProxy
With this method we will now explicitly track all the editor proxies
that are added to the host's workspace. This is symmetric to
`removeEditorProxy`, whose job is to signal when an editor is removed
from the host's workspace.

Note also that the portal delegate methods `addEditorProxy` and
`removeEditorProxy` now take a metadata parameter instead of an entire
editor proxy object.