Remove `addEditorProxy` and `removeEditorProxy`
Instead, every time a new editor proxy is created we will broadcast its
metadata to all the guests. Removing an editor proxy from the portal is
not needed either because we already notify guests when an editor proxy
is disposed.

With this commit we are also introducing two new methods:

* `didChangeEditorProxies` on guest portal delegates, which will be
called every time an editor proxy is created or destroyed.
* `getEditorProxiesMetadata` on Portal, which will return an array
containing the metadata for all the editor proxies that are still alive
in the portal.