Identify sites with no active editor proxy when relaying position info
When notifying delegates of active site positions, include information
for *all* sites. If a site has no active editor proxy (e.g., the site
has given focus to an item outside of the portal, like the "Settings"
tab in Atom), provide a null value for the site's editor proxy and
position.