Add ability for hosts to broadcast buffer URI changes to guests
Change the schema for the `BufferProxyUpdate` message to support
different types of updates. In addition to being able to relay text
operations, the message can now also relay changes to the buffer proxy's
URI.

For the host, the consumers of teletype-client (e.g., the teletype
package for Atom) can use `setURI(uri)` on the `BufferProxy` to inform
guests that the URI has changed. For guests, teletype-client will invoke
`didChangeURI(uri)` on the buffer delegate to inform guests of the
updated URI.

Co-authored-by: Antonio Scandurra <as-cii@github.com>