Ignore leave notification if we don't have info for the given peer
This change is an attempt to prevent the issue reported in
https://github.com/atom/teletype/issues/233#issuecomment-369276052.

We suspect that the root cause is some sort of race condition that
occurs during the initial peer connection process in certain network
conditions. We've attempted to write a test to reproduce the issue, but
we've so far been unsuccessful in doing so. In reviewing the stack trace
in https://github.com/atom/teletype/issues/233#issuecomment-369276052,
we think there's a decent chance that this change will resolve the
issue. For now, we'll ship this change and ask people to let us know if
they continue to experience the bug.

Co-authored-by: Antonio Scandurra <as-cii@github.com>