Don't close RTCPeerConnection if it's already closed
https://github.com/atom/teletype-client/pull/12 added logic to check the
iceConnectionState before attempting to close an RTCPeerConnection. The
issue reported in https://github.com/atom/teletype/issues/368 seems to
indicate that the RTCPeerConnection may be closed even though the
iceConnectionState [1] is not equal to 'closed'. (Perhaps the
iceConnectionState is equal to 'failed' in these scenarios?)

The latest docs seem to indicate that we should instead check the
overall connectionState [2] instead of the iceConnectionState. With the
change in this commit, if the connectionState is not yet 'closed', we
attempt to close the connection.

[1] https://developer.mozilla.org/en-US/docs/Web/API/RTCPeerConnection/iceConnectionState
[2] https://developer.mozilla.org/en-US/docs/Web/API/RTCPeerConnection/connectionState