Check signalingState instead of connectionState
While the docs seem to indicate that connectionState is now the
preferred way to check the state of the connection [1], the
implementation we're using doesn't yet seem to have incorporated those
changes to the WebRTC spec. See the test failure in
https://travis-ci.org/atom/teletype-client/builds/372113418. Therefore,
this commit drops back to using the signalingState instead.

[1] https://developer.mozilla.org/en-US/docs/Web/API/RTCPeerConnection/connectionState