Lock webrtc-adapter to 6.1.x
Upgrading from webrtc-adapter 6.1.5 to 6.2.1 breaks the build similar to
the problem reported in https://github.com/atom/teletype/issues/382. The
package-lock.json file already uses webrtc-adapter 6.1.5 👌, but when
apm installs the teletype package, webrtc-adapter `^6.1.5` gets resolved
to webrtc-adapter `6.x`, for the highest value of x. As a result, this
currently resolves to 6.2.1, which causes the problems referenced above.
By depending on  to webrtc-adapter 6.1.y, we resolve the issue. 😅

[1] https://travis-ci.org/atom/teletype-client/jobs/385265540#L1006